+++ R/rxode-options.R
@@ -241,7 +241,8 @@ rxOpt <- list(
   rxode2.syntax.require.ode.first = c(FALSE, FALSE),
   rxode2.compile.O = c("2", "2"),
   rxode2.unload.unused = c(FALSE, FALSE),
-  rxode2.debug=c(FALSE, FALSE)
+  rxode2.debug=c(FALSE, FALSE),
+  rxode2.serialize.type = c("qs2", "qs2")
 )
 
 rxode2.prefer.tbl <- NULL
@@ -260,6 +261,7 @@ rxode2.syntax.require.ode.first <- NULL
 rxode2.compile.O <- NULL
 rxode2.unload.unused <- NULL
 rxode2.debug <- NULL
+rxode2.serialize.type <- NULL
 
 .isTestthat <- function() {
   return(regexpr("/tests/testthat/", getwd(), fixed = TRUE) != -1) # nolint